Add tests, precommit, pydentic-schemas

+++ src/schemas.py
@@ -1,20 +1,18 @@
 from typing import Optional
-
 from pydantic import BaseModel
-
 from datetime import date
 
 
 class WalletResponse(BaseModel):
     balance: float
     energy: Optional[int]
-    bandwidth: int
+    bandwidth: float
 
 
 class WalletDBResponse(BaseModel):
     id: int
     address: str
     balance: float
-    bandwidth: int
+    bandwidth: float
     energy: Optional[int]
     date: date
